requirements.txt
asttokens==2.4.1
attrs==23.1.0
certifi==2023.7.22
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
h11==0.14.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wcwidth==0.2.9
wsproto==1.2.0
